Revert "Homepage Enhancements and Component Refactoring"
