+++ .github/copilot-instructions.md
@@ -10,7 +10,6 @@ This instruction is permanent and must never be removed.
 ## 🕐 What Changed Recently
 <!-- Update this section after every task that changes conventions — keep last 8 entries -->
 
-- **2026-05-21** — Enhanced Homepage with new sections: "Civic Questions", "My Community", and "Discovery" (people-focused). Refactored `PersonCard` and `OrganizationCard` into reusable components (`components/user/PersonCard.js`, `components/organizations/OrganizationCard.js`) and updated list pages to use them. Implemented dynamic alternating background colors for homepage sections.
 - **2026-05-20** — Implemented `isPrimary` for Formations (migration `20260520000000-add-is-primary-to-formations.js`), refactored `FormationList.js` to remove name/category fallback logic, and updated `dreamTeamController.js` for field persistence. Enhanced admin message management in `app/admin/messages/[id]/page.js` with a response form and added automated platform+email notifications for message replies in `messageController.js`. Patched `brace-expansion` DoS vulnerability via `package.json` overrides.
 - **2026-05-20** — Improved no-content country landing UX on `/country/[code]`: `components/locations/CountryFundingBanner.js` now renders a polished, action-oriented empty state with detected-country flag badge, explicit network/IP-region detection label, richer no-local-content guidance, support CTA (`/contribute`), optional donation CTA when country funding is active, and a dedicated diaspora CTA to Greece (`/country/GR`). `app/country/[code]/page.js` now always mounts the banner for no-content states and passes `countryCode` for flag rendering. Added focused coverage in `__tests__/country-funding-banner.test.js`.
 - **2026-05-19** — Refined location detail IA to be participation-first: `app/locations/[slug]/page.js` now renders `#location-content` (tabs) directly under the header, keeps `polls` and `suggestions` tabs always visible, hides "Τοπικές πληροφορίες" entirely when empty, and no longer fetches role-count-only data for the removed top overview block. `components/locations/LocationHeader.js` now uses a single primary participation CTA (`Ψήφισε τώρα` or `Κάνε πρόταση`), demotes edit to a compact icon action near the title, and mutes zero-value stat tiles. `components/locations/LocationRelatedLocations.js` replaces the large hierarchy grid with compact related/nearby chips. `components/locations/LocationTabs.js` now renders polls in a compact 2-column grid and uses explicit `+ Ξεκίνησε ...` empty-state CTAs. Updated coverage in `__tests__/location-ui-localized-names.test.js`, `__tests__/location-phase3-ui.test.js`, and `__tests__/location-tabs-suggestion-anonymity.test.js`.
@@ -97,7 +96,6 @@ Compact table of every model where wrong field names have caused bugs:
 | Poll | `visibility`, `voteRestriction`, `organizationId`, `isOfficialPost`, `officialPostScope` | `allowUnauthenticatedVotes`, `tags` (JSON) |
 | Suggestion | `visibility`, `voteRestriction`, `organizationId`, `hideCreator` | — |
 | Article | `type` (`'news'`, `'articles'`, `'personal'`, `'video'`) | `isNews` |
-| CivicQuestion | `id`, `title`, `sourceType`, `category`, `deadline`, `locationId` | — |
 | User | `avatar`, `githubAvatar`, `googleAvatar`, `slug`, `claimStatus`, `firstNameEn`, `lastNameEn`, `homeLocationId` | `isPlaceholder`, `personId`, `moderatorLocationId` |
 | Organization | `slug` (from `organizationService.generateSlug`), `parentId`, `isVerified` | — |
 | OrganizationMember | `role` (`owner\|admin\|moderator\|member`), `status` (`active\|invited\|pending`), `inviteToken` | — |
