+++ app/organizations/page.js
@@ -10,7 +10,6 @@ import SkeletonLoader from '@/components/ui/SkeletonLoader';
 import AlertMessage from '@/components/ui/AlertMessage';
 import EmptyState from '@/components/ui/EmptyState';
 import LoadMoreTrigger from '@/components/ui/LoadMoreTrigger';
-import OrganizationCard from '@/components/organizations/OrganizationCard';
 
 import { useAuth } from '@/lib/auth-context';
 
@@ -18,6 +17,40 @@ import organizationTypesConfig from '@/config/organizationTypes.json';
 
 const ORGANIZATION_TYPES = organizationTypesConfig.types;
 
+function OrganizationTypeBadge({ type, t }) {
+  return (
+    <span className="inline-flex items-center px-2 py-0.5 rounded text-xs font-medium bg-blue-100 text-blue-700">
+      {t(`type_${type}`)}
+    </span>
+  );
+}
+
+function OrganizationCard({ organization, t }) {
+  return (
+    <Link href={`/organizations/${organization.slug}`} className="block bg-white rounded-xl shadow-sm border border-gray-200 hover:shadow-md transition-shadow overflow-hidden">
+      <div className="p-5">
+        <div className="flex items-start gap-4">
+          {organization.logo ? (
+            <img src={organization.logo} alt={organization.name} className="h-14 w-14 rounded-lg object-cover border border-gray-200 flex-shrink-0" />
+          ) : (
+            <div className="h-14 w-14 rounded-lg bg-gray-100 border border-gray-200 flex items-center justify-center flex-shrink-0">
+              <BuildingOffice2Icon className="h-7 w-7 text-gray-400" />
+            </div>
+          )}
+          <div className="min-w-0 flex-1">
+            <div className="flex items-center gap-2 flex-wrap">
+              <h2 className="text-base font-semibold text-gray-900 truncate">{organization.name}</h2>
+              <OrganizationTypeBadge type={organization.type} t={t} />
+            </div>
+            {organization.description && (
+              <p className="mt-2 text-sm text-gray-600 line-clamp-2">{organization.description}</p>
+            )}
+          </div>
+        </div>
+      </div>
+    </Link>
+  );
+}
 
 export default function OrganizationsPage() {
   const t = useTranslations('organizations');
@@ -88,7 +121,7 @@ export default function OrganizationsPage() {
         {!error && organizations.length > 0 && (
           <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4">
             {organizations.map((organization) => (
-              <OrganizationCard key={organization.id} organization={organization} />
+              <OrganizationCard key={organization.id} organization={organization} t={t} />
             ))}
           </div>
         )}
